move priority list to json and inline props
