--- components/commission/types.ts
@@ -1,11 +1,0 @@
-export interface CommissionInfoProps {
-  fileName: string
-  PublishDate: string
-  Creator: string
-  Character: string
-  Twitter?: string
-  Pixiv?: string
-  Fantia?: string
-  Skeb?: string
-  Featured: boolean
-}
