--- data/PriorityList.ts
@@ -1,4 +1,0 @@
-export const priorityList: { [key: string]: number } = {
-  Laxy: 99,
-  Lowrise: 100,
-}
