--- data/commission/Alba_sera.ts
@@ -1,20 +1,0 @@
-import { Props } from '#data/types'
-
-export const Sera: Props[] = [
-  {
-    fileName: '20240211_hmng',
-    Character: 'sera',
-    Featured: false,
-    Twitter: 'https://twitter.com/Aozaki__/status/1756552462604132506',
-    Pixiv: '',
-    Skeb: 'https://skeb.jp/@hamident83hami/works/59',
-  },
-  {
-    fileName: '20240305_Lowrise',
-    Character: 'sera',
-    Featured: false,
-    Twitter: 'https://twitter.com/L0WR15E/status/1764969104916848666',
-    Pixiv: '',
-    Skeb: 'https://skeb.jp/@L0WR15E/works/19',
-  },
-]
