--- data/commission/Laxy.ts
@@ -1,44 +1,0 @@
-import { Props } from '#data/types'
-
-export const Laxy: Props[] = [
-  {
-    fileName: '20210424_Laxy',
-    Character: 'tkmt',
-    Featured: true,
-    Twitter: 'https://twitter.com/laxyiii/status/1385985122332155908',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210531_Laxy',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/laxyiii/status/1402415053567975424',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20211018_Laxy',
-    Character: 'tkmt',
-    Featured: true,
-    Twitter: 'https://twitter.com/laxyiii/status/1451854024609714183',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20220113_Laxy',
-    Character: 'tkmt',
-    Featured: true,
-    Twitter: 'https://twitter.com/laxyiii/status/1481568202060709892',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20220628_Laxy',
-    Character: 'tkmt',
-    Featured: true,
-    Twitter: 'https://twitter.com/laxyiii/status/1541634796887019522',
-    Pixiv: '',
-    Skeb: '',
-  },
-]
