--- data/commission/Lucia.ts
@@ -1,21 +1,0 @@
-import { Props } from '#data/types'
-
-export const Lucia: Props[] = [
-  /**
-  {
-    fileName: '20230701_淵゛',
-    Character: 'lucia',
-    Featured: false,
-    Twitter: 'https://twitter.com/Qooo003/status/1675005730800435201',
-    Pixiv: '',
-    Skeb: 'https://skeb.jp/@Qooo003/works/40',
-  },*/
-  {
-    fileName: '20230703_Lowrise',
-    Character: 'lucia',
-    Featured: false,
-    Twitter: 'https://twitter.com/L0WR15E/status/1675836797627006976',
-    Pixiv: '',
-    Skeb: 'https://skeb.jp/@L0WR15E/works/11',
-  },
-]
