--- data/commission/Stale/Nishe.ts
@@ -1,20 +1,0 @@
-import { Props } from '#data/types'
-
-export const Nishe: Props[] = [
-  {
-    fileName: '20221022_RENXZD',
-    Character: 'nishe',
-    Featured: false,
-    Twitter: 'https://twitter.com/ASD03853116/status/1587269529586933762',
-    Pixiv: 'https://www.pixiv.net/artworks/102419272',
-    Skeb: '',
-  },
-  {
-    fileName: '20221030_RENXZD',
-    Character: 'nishe',
-    Featured: false,
-    Twitter: 'https://twitter.com/ASD03853116/status/1584420524641292289',
-    Pixiv: '',
-    Skeb: '',
-  },
-]
