--- data/commission/Stale/Tkmt.ts
@@ -1,108 +1,0 @@
-import { Props } from '#data/types'
-
-export const Tkmt: Props[] = [
-  {
-    fileName: '20211018_四ノ宮しの',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/sinosino141',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210708_四ノ宮しの',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/sinosino141',
-    Pixiv: '',
-    Skeb: 'https://skeb.jp/@sinosino/works/131',
-  },
-  {
-    fileName: '20220321_四ノ宮しの',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/sinosino141/status/1505863501885042691',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210902_四ノ宮しの',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/sinosino141/status/1433446939106807813',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210516_四ノ宮しの',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/sinosino141/status/1393899030342782977',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210525_のんびり',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/nonbi_re/status/1397497316060270600',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210815_のんびり',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/nonbi_re/status/1426826402473517059',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210331_ALETH',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/riva_poul/status/1377187662084341760',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210523_ALETH',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/riva_poul/status/1396390446390300674',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210428_楠ハルイ',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/hr_x9_/status/1387324035655036930',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210217_金沢',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/animarcat',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20210316_朱耶式',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/akaya_siki/status/1371584249745186817',
-    Pixiv: '',
-    Skeb: '',
-  },
-  {
-    fileName: '20220403_雨音',
-    Character: 'tkmt',
-    Featured: false,
-    Twitter: 'https://twitter.com/walabemoti2/status/1510469282572156931',
-    Pixiv: '',
-    Skeb: '',
-  },
-]
