+++ data/priorityList.json
@@ -1,0 +1,4 @@
+{
+  "Laxy": 99,
+  "Lowrise": 100
+}
