--- data/types.ts
@@ -1,8 +1,0 @@
-export interface Props {
-  fileName: string
-  Character: string
-  Featured: boolean
-  Twitter: string
-  Pixiv: string
-  Skeb: string
-}
